Move the custom exceptions to the exceptions directory
