+++ src/exceptions/ArgumentError.js
@@ -1,0 +1,11 @@
+var util = require('util');
+
+
+var ArgumentError = function(message){
+  this.message = message;
+};
+
+util.inherits(ArgumentError, Error);
+
+
+module.exports = ArgumentError;
