+++ src/exceptions/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  APIError: require('./APIError.js'),
+  ArgumentError: require('./ArgumentError.js')
+};
